refactor(pty): #738 session.rs god-file を sub-module 分割し lock helper / VecDeque 定数を集約
`src-tauri/src/pty/session.rs` (約1800行) の god-file を `pty/session/` 配下の
サブモジュールに分割する純粋なリファクタ。PTY の spawn/inject/resize/exit 挙動・
Windows パス解決ロジック・Mutex poison 時挙動は一切変えていない。

- pty/session.rs を pty/session/ 配下に分割:
  - handle.rs        — SessionHandle + Drop + UserWriteOutcome / TerminalExitInfo
  - spawn.rs         — SpawnOptions / spawn_session / cwd・コマンドパス解決 / spawn メトリクス
  - env_allowlist.rs — 親プロセス環境変数の継承 allowlist (should_inherit_env)
  - windows_resolve.rs — Windows 専用パス解決 (#[cfg(windows)])
  - injecting_guard.rs — injecting フラグの RAII guard (InjectingGuard)
  - lock.rs          — Mutex lock helper
  公開 API は session/mod.rs の re-export で `crate::pty::session::*` 互換を維持。
  純粋ロジック / 公開 API 系テストは pty/tests/session_{env,spawn,windows}.rs へ移設。

- Mutex lock helper 共通化: LockResult<T> エイリアス + lock_poisoned!(name) macro を
  pty/session/lock.rs に切り出し、SessionHandle の 4 個の `.lock().map_err(...)` 重複
  (writer / master / killer / write_budget) を 1 macro 呼び出しに集約。エラー文言は
  従来どおり "<name> lock poisoned: {e}" のまま。

- VecDeque 上限定数集約: team_hub/protocol/consts.rs に MAX_TEAM_REPORTS /
  MAX_WORKER_REPORTS / MAX_HANDOFF_EVENTS / MAX_NEXT_ACTIONS を集約し、report.rs /
  send.rs / update_task.rs / state/persistence.rs に散在していた inline リテラル
  (50 / 20) を全て定数参照に置換。

Closes #738
